wrap: parse env-if-host options as filenames instead of strings
If we parse the "env-if-host" options as "G_OPTION_ARG_STRING_ARRAY",
Glib assumes them to be encoded with the current locale. This can be an
issue when pressure-vessel needs to run the main game, because Steam
will set "LC_ALL=C", but the "env-if-host" might still contain
characters that are not ASCII.
To solve this we use "G_OPTION_ARG_FILENAME_ARRAY" that treats the
options as opaque byte-blobs.

Fixes: #52
Fixes: steam-runtime#349

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>