check-vulkan: Convert the file from C++ to C
Our entire codebase is in C with the only exception for check-vulkan and
check-gl.

They were coded in C++ because they were based on pre-existing codes.

Converting check-vulkan in C allows us to conform it to the rest of our
codebase and also it will be easier to expand its functionalities, like
for example as outlined in #50.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>