Backport the new json_from_string() function
JSON-GLib 1.2.0 introduced a new function called `json_from_string()`.

It is an handy function that allows us to obtain a jsonNode with just a
single operation.

This function had to be backported, and modified a little, because we
are
targeting Scout that has an older JSON-GLib version that misses both
`json_from_string()` and `json_parser_steal_root()` (that was used in
`json_from_string()`).

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>