system-info: Print diagnostic messages as arrays of lines
A single line with escaped newlines is unnecessarily hard to read.

This repurposes code that we already had for printing a udev uevent
pseudo-file in this way.

We retain the ability to parse old s-r-s-i output in which the
diagnostic messages were a string: the intended meaning is fairly
obvious.

Signed-off-by: Simon McVittie <smcv@collabora.com>