pressure-vessel: Accept runtime with or without ./files
Previously, if we were using a Flatpak-style runtime, the argument
to --runtime= had to be its ./files subdirectory. Now we accept the
top-level directory as an alternative.

This will make it more straightforward for pressure-vessel-wrap to
take responsibility for unpacking and deploying runtimes that are
shipped as a tarball, which is currently done in the SteamLinuxRuntime
shell scripts, resulting in error handling and locking being less
robust.

Signed-off-by: Simon McVittie <smcv@collabora.com>