runtime: improve libraries binding speed
Instead of calling bwrap every time we need to bind a library, or the
dependencies of a library, we group them together to speed up the whole
process.

In `bind_icd()`, to decide whether to use a numbered subdir or not, we
check if a symlink for the same library already exists. It also compares
the directory size before and after the binding to check if we were
actually able to capture it.
For this reason we decided to not group together the ICDs/Layers, to
avoid having side effects like in the previous commit
"runtime: improve ICDs binding speed" and to avoid an overly complicated
hard to follow implementation. Instead we group together only the
libraries that need to be binded in `arch->libdir_in_current_namespace`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>