pv-wrap: Don't try so hard to use host graphics by default
For historical reasons, we assume that /run/host is a relatively
complete view of the host system, like it would be inside Flatpak.
However, in a systemd-nspawn container, /run/host contains other
things but not usually /usr or /etc.

Previously, the default was exactly equivalent to the old
PRESSURE_VESSEL_HOST_GRAPHICS=1, which in turn is equivalent to either
PRESSURE_VESSEL_GRAPHICS_PROVIDER=/run/host (if we appear to be in a
container) or PRESSURE_VESSEL_GRAPHICS_PROVIDER=/ (if not).

Now, the default is basically PRESSURE_VESSEL_GRAPHICS_PROVIDER=/, to
maximize the chance that if Steam games can run in the older
LD_LIBRARY_PATH-based Steam Runtime, they'll also run in our containers.

Resolves: https://github.com/ValveSoftware/steam-runtime/issues/367
Signed-off-by: Simon McVittie <smcv@collabora.com>