pressure-vessel: Treat LD_AUDIT the same as LD_PRELOAD
The Flathub Steam app loads a LD_AUDIT module to ignore bundled libraries
in games where they cause problems. We want to treat that essentially
the same as LD_PRELOAD, at least until we take on its functionality.

While we're changing this, also change the misleading name of the
option. It was called --host-ld-preload, but it actually took the
LD_PRELOAD module from the current execution environment, not the host.

Signed-off-by: Simon McVittie <smcv@collabora.com>