runtime: Bind the drirc settings from the provider
If we are using the provider graphics stack, including mesa, we should
also bind the drirc settings. Otherwise we would end up using the older
settings from the runtime.

The default configuration file is expected to be located under
`/usr/share/drirc.d`, with an additional system-wide configuration file
called `/etc/drirc` and a per-user configuration `$HOME/.drirc`.

The only one that has been left out by this commit is the one that comes
from `$HOME`, because we normally already share the whole home
directory, so we expect it to be already available in the container.

Fixes: #60

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>