wrap: Put game-specific $HOME on bwrap command-line before exports
This is how Flatpak behaves, and it's necessary to prevent the
game-specific $HOME from being mounted over the top of some of the
exports, preventing (for example) ~/.steam from being exposed to the
container.

We only do this for $HOME rather than completely reverting 3a8e939d,
so that it won't be a regression for a symlink at /var/tmp.

If the logical and physical paths to the real $HOME differ, we now also
mount the game-specific $HOME onto the real $HOME's physical path, and
replicate the logical path's symlinks in the container. This matches how
Flatpak works, and avoids a similar regression.

Fixes: 3a8e939d "pv-wrap: Delay --dir, etc. until after FlatpakExports are processed"
Resolves: https://gitlab.steamos.cloud/steamrt/steam-runtime-tools/-/issues/63
Signed-off-by: Simon McVittie <smcv@collabora.com>