runtime: Create library aliases in overrides
If a library has different SONAMEs, called aliases in
"steam-runtime-abi.json", it might lead to errors when a game tries to
load a SONAME that is not the "real" library SONAME (the DT_SONAME that
objdump shows). Because `ld.so.cache` will not consider the aliases when
searching for the right match.

To workaround this problem we create a directory called "aliases" that
contains symlinks of the libraries aliases that we know about.

By adding these directories, one for every multiarch, in the
`LD_LIBRARY_PATH`, we ensure that even games that were linked to an
alias will be able to find the library they were looking for.

Helps: https://github.com/ValveSoftware/steam-runtime/issues/246
Fixes: T26816

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>