runtime: Use "/usr/share" as a fallback lib data directory
If the host system stores the shared libraries in a directory that is
not what we expect, we try as a last resort `/usr/share`, in an effort
to increase the number of OSs that we are able to cover.

For example this should help Exherbo to find the right libdrm and
drirc.d directories.

Fixes: #68

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>