wrap: Use a tmpfs home as a fallback when unsharing the home dir
If we don't want to share the host home directory we expect to have
either --home, --freedesktop-app-id, --steam-app-id or $SteamAppId.

There are cases where this might not happen. For example if Steam has
been launched with the PRESSURE_VESSEL_SHARE_HOME=0 env variable and we
try to run "Help->System Information", the variable $SteamAppId will not
be set and the container creation will fail.

To avoid this issue we can use a tmpfs for the home directory when we
are in batch mode.
In this way the tmpfs will not be used if we need to run an actual game,
because we want to retain the home directory on exit.

Fixes: #66

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>