inspect-library: Add a line-based octal output format
Instead of trying to format JSON using only glibc, this just prints
key=value lines, some of which can be repeated. The values always have
control characters and backslashes escaped as C-style octal sequences
such as \040 for space, similar to the encoding of /etc/fstab. This
allows non-UTF-8 bytestrings to be represented.

Signed-off-by: Simon McVittie <smcv@collabora.com>