steam-remote: Add a stub steam executable
This stub steam executable tries to directly pass the given commands,
e.g. "steam://nav/downloads", to the running Steam client, if any.

This helps games that tries to run "steam ${command}" because when we
are in a LD_LIBRARY_PATH runtime, "steam" might not be pointing to the
same version that is currently running. Or if we are in a container,
there is no "steam" executable in the PATH at all.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>