urlopen: Add a more clever xdg-open executable
When inside a Steam Runtime container we currently make `xdg-open`
points to the `flatpak-xdg-utils` implementation that relies on
`xdg-desktop-portal`. But if an user doesn't have a working
`xdg-desktop-portal` it will fail.

Additionally the `steam://` URLs might end up opening another instance
of the Steam client, even if one instance was already running. This can
happen if more than one version are installed, e.g. the Steam client
from the distro repositories and the Flatpak version.

With this alternative implementation we try a more clever approach that
should workaround the `xdg-open` shortcomings that we experienced in our
Steam related use cases.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>
[smcv: Install all /usr/bin/steam-runtime-* executables]
Signed-off-by: Simon McVittie <smcv@collabora.com>