wrap: Send a more clear warning when attempting to bind /usr/* dirs
pressure-vessel doesn't support having the Steam Library under `/usr`.

When this happens, pressure-vessel fails without a clear indication of
what went wrong. For this reason we add a warning message to point out
that pressure-vessel was asked to bind a directory that will not appear
in the resulting container.

Helps: https://github.com/ValveSoftware/steam-runtime/issues/288
Helps: #30

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>