runtime: Search for executables using the provider instead of the host
When we are in a Flatpak environment we want to pick "localedef",
"locale" and "ldconfig" from the Flatpak runtime instead of the ones
provided by Scout/Soldier.

By searching in the "provider_in_current_namespace" we can ensure to
create a subsandbox where these executables are symlinked to the
"/run/parent" counterpart.

Fixes: T27438

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>