pv-wrap: Remove support for escaping from a Flatpak container
We did this as a very early prototype, but it never fully worked, and
required configuring Flatpak specially to allow Steam to escape from the
sandbox (which obviously undermines Flatpak's intended sandboxing).

Now that we've implemented the features we need in Flatpak's
sub-sandboxing code, we have a better way to handle this, and can
simplify the code by not trying to have it both ways.

Signed-off-by: Simon McVittie <smcv@collabora.com>