Add support to detect the expansion of $LIB and $PLATFORM
When loading VDPAU modules we rely on $PLATFORM, trying to support all
its known possible expansions.

However there is always the possibility that we are currently missing
some of them or that new ones will be added in the future.

For this reason we try to detect the expansion of $PLATFORM, and $LIB,
and print its value in the report. In this way it will be easier to
notice unusual and/or new expansions.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>