pv-runtime: Avoid opening libraries to determine their SONAME
This is surprisingly expensive to do, particularly with a cold cache.
If we can avoid this, then deleting unwanted libraries from the mutable
copy of the runtime becomes a lot faster.

With the disk cache cleared and running on slow hardware, this cuts
down deletion of the unwanted libraries from 2-4 seconds to basically
instantaneous.

Signed-off-by: Simon McVittie <smcv@collabora.com>