pv-runtime: Convert the graphics provider into an object
This encapsulates both the PROVIDER_GRAPHICS_STACK flag and the
associated paths: if the object is null then the paths are meaningless,
and if the object is non-null then they are meaningful.

Making this an immutable "value object" also means we can share it
between threads, unlike PvRuntime, which has state. This could become
important if we want to make graphics driver enumeration multi-threaded
to speed up pressure-vessel.

Signed-off-by: Simon McVittie <smcv@collabora.com>