Never redirect stdout to the log file for setup commands
If we redirect the stdout to a file, Origin starts to complain that it
is not able to find the game it was supposed to launch.
This seems to be caused by Steam that does something like this:
```
installation_path = subprocess.check_output(".../_v2-entry-point
--verb=run -- .../proton run .../iscriptevaluator.exe
--print-installation-path")
```

And if stdout is redirected it doesn't work and complains with:
```
src/clientdll/installscript_posix.cpp (357) : Assertion Failed:
strCommonRedistForeignInstallPath != ""
```

We already avoided the redirect of stdout when `PRESSURE_VESSEL_BATCH`
was set, but apparently this environment variable is not used when we
are in the setup phase of a game.

Fixes: https://github.com/ValveSoftware/steam-runtime/issues/402

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>