populate-depot: Allow unpacking runtimes into a versioned directory
This allows multiple versions to coexist for local testing, and is also
a step towards avoiding the double disk space consumption of deploying
the runtime from an archive.

The runtime is now unpacked into a versioned directory by default: the
Steam CDN now knows how to handle this efficiently, moving unmodified
files from one runtime to the next. Using a versioned name ensures that
any locally-added files applied by a user will stay where they're put
instead of propagating into updated runtimes.

The versioned directory name contains `_platform_`, for example
`soldier_platform_0.20210428.0`, both for symmetry with the SDK and
so that it will not be deleted by pv-wrap --gc-legacy-runtimes.

Signed-off-by: Simon McVittie <smcv@collabora.com>