wrap: Set the env var to adjust libshared-library-guard configuration path
When inside a Flatpak container, we expect that
`libshared-library-guard.so` looks for its blocklist configuration in
`/app/etc/freedesktop-sdk.ld.so.blockedlist`.

This will not work because we bind `/app` in `/run/parent/app`. For this
reason, when we detect this situation, we set the environment variable
`SHARED_LIBRARY_GUARD_CONFIG` to point `libshared-library-guard` in the
correct path.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>