populate-depot: Make local runtime compatible with --unpack-runtimes
Since commit 42fb977a, run() expects to find the tarball in ./.cache.
If we're populating the depot with a newly-built runtime, as we do in
order to smoke-test new versions of runtimes, then we need to satisfy
that assumption by hardlinking it into .cache unconditionally, the
same way 42fb977a downloads tarballs into .cache unconditionally.

Similarly, since commit bfb0b34d, we might not want to include the
tarball (and its build-ID flag file) in the depot. Only put the tarball
in the depot if wanted.

The pinned_version for a locally-built runtime is None, but we want
to use the same build ID for the locally-built runtime that we put
in the directory name, so we need to fall back to runtime.version
when writing out the ./run and ./run-in-$suite scripts.

Signed-off-by: Simon McVittie <smcv@collabora.com>