Add a way to layer scout onto soldier
This works like Proton 5.13+, with Steam Runtime 2 'soldier' as the
"lower" layer; but instead of Proton, the "upper" layer is the traditional
LD_LIBRARY_PATH Steam Runtime. This has compatibility characteristics
similar to running a non-container runtime on a Debian 10 host system,
and in particular can run some games that don't work in a strict scout
container, such as Life Is Strange 2.

Resolves: T18668
Signed-off-by: Simon McVittie <smcv@collabora.com>