resolve-in-sysroot: Correct misleading error messages
At this point, current_path is everything up to but not including fd.
For example, if we are traversing "x/y/z/w" with current_path = "x/y",
then the fds array contains { x, y } and fd is the result of opening z.
We're looking at the result of calling fstatat() on z, so we need to
see "x/y/z" in the error messages, for example if z is a regular file
(therefore "x/y/z/w" cannot possibly exist).

Signed-off-by: Simon McVittie <smcv@collabora.com>