mtree: Parse modification times in a compatible way
time=1.234 has historically meant 1 second + 234 nanoseconds, or what
normal people would write as 1.000000234. Parse it as such, but for now
treat this ambiguous format as an error, unless the fractional part is
exactly ".0" which is unambiguous - we can relax this to a warning
later, if we want to.

libarchive and FreeBSD mtree show mtimes with exactly 9 decimal digits,
which makes them mean the same thing in either interpretation.

Signed-off-by: Simon McVittie <smcv@collabora.com>