mtree: Parse the ignore, nochange and optional flags
The nochange flag is fully implemented here: it suppresses changing the
modification time or permissions.

The optional flag is mostly implemented here: it suppresses the error
that would ordinarily occur if a file with content does not already
exist.

The ignore flag is not yet implemented, only parsed, because it only
really applies when verifying whether a filesystem tree matches a
specification.

Signed-off-by: Simon McVittie <smcv@collabora.com>