CI: Make relocatable builds with both production and debug binaries
The production build uses dpkg packages, built with -O2 -g and stripped.
This is similar to what we currently extract from a scout SDK build;
but if we do it here, we can decouple the pressure-vessel release process
from the scout release process, making it possible to issue a new
pressure-vessel release for use with pre-existing scout/soldier runtimes
much quicker than the time needed to do an entire scout release.

Signed-off-by: Simon McVittie <smcv@collabora.com>