scout-on-soldier: Fail if toolmanifest.vdf has not been reloaded
If the user was previously running a self-contained version of scout,
toolmanifest.vdf will need to be reloaded before we will get run in a
soldier container as required. Implement the entry point path used in
the self-contained version as a script that just pops up a zenity error.
(T28437)

Signed-off-by: Simon McVittie <smcv@collabora.com>