tests: Don't rely on being able to put more than one suite in a depot
The SteamLinuxRuntime scripts have phased out support for populating a
depot directory with a mixture of scout and soldier, because that
doesn't really make sense for a game or a compat tool: each native Linux
game, and each compat tool used to run non-native games, expects to run
in one particular environment (scout or soldier or something else) and
using any other environment would be incorrect.

However, tests/pressure-vessel/containers.py currently expects to work
with a directory that contains both scout and soldier. For now, we'll
download scout and soldier separately, then mash them together into the
same directory.

Signed-off-by: Simon McVittie <smcv@collabora.com>