bwrap: Cope with /dev/shm being a symlink to /run/shm
This can be the case on older sysvinit- or Upstart-based Debian-derived
distributions, such as Ubuntu 14.04 (with the default Upstart init
system) and Debian 9 (with the non-default sysvinit init system).

Under systemd or the versions of sysvinit in Debian >= 10, /dev/shm is
a real tmpfs and /run/shm is a symlink to /dev/shm.

Signed-off-by: Simon McVittie <smcv@collabora.com>