populate-depot: Have an overall version number for the depot as a whole
With some changes coming in via pressure-vessel, some via the container
image and some via the scripts, any or all of which might be updated
in any given container build, it's getting a bit confusing. If we assign
an overall version number to the whole container, we can talk about a
particular depot build unambiguously.

Signed-off-by: Simon McVittie <smcv@collabora.com>