populate-depot: Use latest scout build for LD_LIBRARY_PATH runtime
If we're building a SteamLinuxRuntime_soldier depot and we want to test
launching it from a LD_LIBRARY_PATH runtime (as Steam would), there is
no point in trying to match the version number of scout to a version
number of soldier. Worst case it'll just fail, and best case it'll use
an arbitrary scout build that happens to have been done on the same day
as the soldier build.

This is particularly problematic right now because the latest build of
soldier happens to be 0.20210618.0. scout build 0.20210618.0 existed,
but had a broken steam-runtime.tar.xz, because the setup.sh script is
not executable.

Signed-off-by: Simon McVittie <smcv@collabora.com>