scout-on-soldier: Do our own pinning instead of running setup.sh
Because scout and soldier are relatively fixed targets, and we know that
soldier is strictly newer than scout, we don't need to go to all this
effort to figure out which libraries are newer if they exist in both
scout and soldier: the answer is always "the one in soldier".

This means the only library we need to pin is the single library that
is hard-coded to be pinned, namely libcurl.so.4, for which the ABI broke
between scout and soldier.

Signed-off-by: Simon McVittie <smcv@collabora.com>