pressure-vessel: Remove the concept of "locking" environment variables
This was only necessary because we were reusing a single container
across multiple entry-point invocations, and expecting "most" arbitrary
environment variables from each new invocation to be taken into account
for commands running in the container, which meant that we needed to
keep track of which environment variables had to be exceptions to that
rule for technical reasons. Now that we're no longer injecting multiple
commands into the same container like that, we don't need this
complexity.

Signed-off-by: Simon McVittie <smcv@collabora.com>