tests/scout.sh: Use latest-steam-client-public-beta
The latest-container-runtime-public-beta alias no longer exists, because
the latest container runtime public beta for scout doesn't contain its
own copy of scout any more, making the alias misleading. Instead, it's
the "layered" runtime previously seen in the scout_layered_slim branch,
where we use a Steam Runtime 2 'soldier' container and then layer the
scout LD_LIBRARY_PATH runtime included with Steam onto it, for an
overall result similar to using the traditional LD_LIBRARY_PATH
(non-container) runtime on a Debian 10 system. This provides better
compatibility with games that were compiled in a non-pure scout
environment, such as Life Is Strange 2.

The closest equivalent is the latest-steam-client-public-beta alias,
which represents the version of the LD_LIBRARY_PATH runtime that is
included in the public beta branch of the Steam client, so move over
to that when producing non-"layered" scout runtime builds.

Signed-off-by: Simon McVittie <smcv@collabora.com>