tests: Use more default settings
The download URL defaults to the public download URL and the version of
pressure-vessel defaults to the latest standalone release, so we don't
usually need to specify those two explicitly.

In the case of PRESSURE_VESSEL_DOWNLOAD_URL, if it has been explicitly
set, we still want to pass it in: the official CI does this to bypass
the CDN that implements repo.steampowered.com and go directly to the
origin server.

Signed-off-by: Simon McVittie <smcv@collabora.com>