runtime: Share /usr/share/nvidia if we have at least one libGLX_nvidia
If we have libGLX_nvidia.so.0 for *any* architecture - even if we are
msising some instances - then we still want to share /usr/share/nvidia
with the container.

Because we always use libGLX_nvidia.so.0 from the graphics stack
provider and do not have a concept of whether it is older or newer,
and we do not expect our runtime to have a copy of libGLX_nvidia.so.0,
we do not need to worry about giving the runtime's library an
incompatible version of the data files from the provider.

Signed-off-by: Simon McVittie <smcv@collabora.com>