_v2-entry-point: Remove support for not relaunching the container
This effectively hard-codes PRESSURE_VESSEL_RELAUNCH_CONTAINER to 1:
if we are in "session mode" with a STEAM_COMPAT_SESSION_ID set, then
we will still share a container started by _start-container-in-background
between one or more setup commands, but then we will terminate that
container and start again for the actual game.

We need to do this if we want the actual game to be able to take its
user-specified launch options into account, which is why it has been the
default since January.

Part of T25709.

Signed-off-by: Simon McVittie <smcv@collabora.com>