pv-adverb: Allow setting the architecture of a module in the CLI
This will allow pv-wrap to translate a host `LD_PRELOAD` item like
`/usr/$LIB/libMangoHud.so` into a pair of `--ld-preload` options, each
looking like `--ld-preload=/usr/lib32/libMangoHud.so:abi=i386-linux-gnu`.

The only character that cannot appear in a `LD_PRELOAD` path entry is `:`,
so we're free to use that as our representation for extra options.
Instead of just appending the ABI tuple, use a key/value notation so
we can extend later if we need to.

Signed-off-by: Simon McVittie <smcv@collabora.com>